refactor: add enrichStatus
